requirements.txt
# Knowledge Distillation Requirements
# ====================================

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# ONNX support
onnx>=1.14.0
onnxruntime-gpu>=1.15.0  # Use onnxruntime for CPU-only systems

# TensorFlow for TFLite export
tensorflow>=2.14.0,<2.17.0  # TF 2.14-2.16 for Android compatibility
onnx-tf>=1.10.0

# ONNX simplification (optional but recommended)
onnxsim>=0.4.0

# Data handling
numpy>=1.24.0
Pillow>=9.0.0

# Configuration
PyYAML>=6.0

# Progress bars and logging
tqdm>=4.65.0

# Visualization and evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics and evaluation
scikit-learn>=1.3.0
seaborn>=0.12.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
